src/requirements.txt
# KnowlEdge 个性化知识引擎依赖
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
jinja2>=3.1.2
sse-starlette>=1.3.4
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.28.2
arxiv>=1.4.7
PyPDF2>=3.0.1
python-docx>=0.8.11
pandas>=2.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
transformers>=4.31.0
torch>=2.0.0
numpy>=1.24.2
websockets>=11.0.3
aiofiles>=23.1.0
scholarly>=1.7.11
pydantic>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.2 